Refactor Gherkin features: put all features related with sessions in a file. Because the `steps_app_language.go` is very big.
