+++ testhelpers/steps_app_language.go
@@ -416,57 +416,6 @@ func (ctx *TestContext) addThread(item, participant, helperGroup, status, messag
 	})
 }
 
-// addSession adds a session in database.
-func (ctx *TestContext) addSession(session, user, refreshToken string) {
-	sessionID := ctx.getReference(session)
-	userID := ctx.getReference(user)
-
-	ctx.addInDatabase("sessions", strconv.FormatInt(sessionID, 10), map[string]interface{}{
-		"session_id":    sessionID,
-		"user_id":       userID,
-		"refresh_token": refreshToken,
-	})
-}
-
-// addAccessToken adds an access token in database.
-func (ctx *TestContext) addAccessToken(session, token, issuedAt, expiresAt string) {
-	sessionID := ctx.getReference(session)
-
-	issuedAtDate, err := time.Parse(utils.DateTimeFormat, issuedAt)
-	if err != nil {
-		panic(err)
-	}
-
-	expiresAtDate, err := time.Parse(utils.DateTimeFormat, expiresAt)
-	if err != nil {
-		panic(err)
-	}
-
-	ctx.addInDatabase("access_tokens", token, map[string]interface{}{
-		"session_id": sessionID,
-		"token":      token,
-		"issued_at":  issuedAtDate,
-		"expires_at": expiresAtDate,
-	})
-}
-
-// IAm Sets the current user.
-func (ctx *TestContext) IAm(name string) error {
-	err := ctx.ThereIsAUser(name)
-	if err != nil {
-		return err
-	}
-
-	err = ctx.IAmUserWithID(ctx.getReference(name))
-	if err != nil {
-		return err
-	}
-
-	ctx.user = name
-
-	return nil
-}
-
 // ThereIsAUser create a user.
 func (ctx *TestContext) ThereIsAUser(name string) error {
 	return ctx.ThereIsAUserWith(getParameterString(map[string]string{
@@ -990,107 +939,4 @@ func (ctx *TestContext) IAmPartOfTheHelperGroupOfTheThread() error {
 // to a given item.
 func (ctx *TestContext) ICanRequestHelpToTheGroupWithIDOnTheItemWithID(group, item string) error {
 	return ctx.UserCanRequestHelpToOnItemWithID(group, ctx.user, item)
-}
-
-// ThereAreTheFollowingSessions create sessions.
-func (ctx *TestContext) ThereAreTheFollowingSessions(sessions *messages.PickleStepArgument_PickleTable) error {
-	for i := 1; i < len(sessions.Rows); i++ {
-		session := ctx.getRowMap(i, sessions)
-
-		ctx.addSession(
-			session["session"],
-			session["user"],
-			session["refresh_token"],
-		)
-	}
-
-	return nil
-}
-
-// ThereAreCountSessionsForUser checks if there are a given number of sessions for a given user.
-func (ctx *TestContext) ThereAreCountSessionsForUser(count int, user string) error {
-	userID := ctx.getReference(user)
-
-	var sessionCount int
-	err := ctx.db.QueryRow("SELECT COUNT(*) as count FROM sessions WHERE user_id = ?", userID).
-		Scan(&sessionCount)
-	if err != nil {
-		return err
-	}
-
-	if sessionCount != count {
-		return fmt.Errorf("expected %d sessions for user %s, got %d", count, user, sessionCount)
-	}
-
-	return nil
-}
-
-func (ctx *TestContext) ThereIsNoSessionID(session string) error {
-	sessionID := ctx.getReference(session)
-
-	var sessionCount int
-	err := ctx.db.QueryRow("SELECT COUNT(*) as count FROM sessions WHERE session_id = ?", sessionID).
-		Scan(&sessionCount)
-	if err != nil {
-		return err
-	}
-
-	if sessionCount > 0 {
-		return fmt.Errorf("there should be no session with ID %d", sessionID)
-	}
-
-	return nil
-}
-
-// ThereAreTheFollowingAccessTokens create access tokens.
-func (ctx *TestContext) ThereAreTheFollowingAccessTokens(accessTokens *messages.PickleStepArgument_PickleTable) error {
-	for i := 1; i < len(accessTokens.Rows); i++ {
-		accessToken := ctx.getRowMap(i, accessTokens)
-
-		ctx.addAccessToken(
-			accessToken["session"],
-			accessToken["token"],
-			accessToken["issued_at"],
-			accessToken["expires_at"],
-		)
-	}
-
-	return nil
-}
-
-// ThereAreCountAccessTokensForUser checks if there are a given number of access tokens for a given user.
-func (ctx *TestContext) ThereAreCountAccessTokensForUser(count int, user string) error {
-	userID := ctx.getReference(user)
-
-	var accessTokensCount int
-	err := ctx.db.QueryRow(`
-		SELECT COUNT(*) as count FROM access_tokens
-			JOIN sessions ON sessions.session_id = access_tokens.session_id
-		 WHERE sessions.user_id = ?`, userID).
-		Scan(&accessTokensCount)
-	if err != nil {
-		return err
-	}
-
-	if accessTokensCount != count {
-		return fmt.Errorf("expected %d access tokens for user %s, got %d", count, user, accessTokensCount)
-	}
-
-	return nil
-}
-
-// ThereIsNoAccessToken checks that an access token doesn't exist.
-func (ctx *TestContext) ThereIsNoAccessToken(accessToken string) error {
-	var accessTokensCount int
-	err := ctx.db.QueryRow("SELECT COUNT(*) as count FROM access_tokens WHERE token = ?", accessToken).
-		Scan(&accessTokensCount)
-	if err != nil {
-		return err
-	}
-
-	if accessTokensCount > 0 {
-		return fmt.Errorf("there should be no access token %s", accessToken)
-	}
-
-	return nil
 }
